requirements.txt
google-generativeai
opencv-python
pyserial
SpeechRecognition
pyttsx3
pyaudio
